Google Workspace: OAuth utils
