+++ extensions/google-workspace/package.json
@@ -92,7 +92,7 @@
   ],
   "dependencies": {
     "@raycast/api": "^1.62.0",
-    "@raycast/utils": "^1.10.1",
+    "@raycast/utils": "^1.12.4",
     "date-fns": "^2.30.0",
     "jwt-decode": "^3.1.2",
     "node-fetch": "^3.3.2"
