--- extensions/google-workspace/src/api/oauth.ts
@@ -1,88 +1,0 @@
-import { OAuth } from "@raycast/api";
-import jwt_decode from "jwt-decode";
-import fetch from "node-fetch";
-
-const clientId = "859594387706-uunbhp90efuesm18epbs0pakuft1m1kt.apps.googleusercontent.com";
-
-export const client = new OAuth.PKCEClient({
-  redirectMethod: OAuth.RedirectMethod.AppURI,
-  providerName: "Google",
-  providerIcon: "google-logo.png",
-  providerId: "google",
-  description: "Connect your Google account",
-});
-
-export async function authorize(): Promise<string> {
-  const tokenSet = await client.getTokens();
-  if (tokenSet?.accessToken) {
-    if (tokenSet.refreshToken && tokenSet.isExpired()) {
-      const tokens = await refreshTokens(tokenSet.refreshToken);
-      await client.setTokens(tokens);
-      return tokens.access_token;
-    }
-    return tokenSet.accessToken;
-  }
-
-  const authRequest = await client.authorizationRequest({
-    endpoint: "https://accounts.google.com/o/oauth2/v2/auth",
-    clientId,
-    scope:
-      "https://www.googleapis.com/auth/userinfo.email https://www.googleapis.com/auth/userinfo.profile https://www.googleapis.com/auth/drive.readonly",
-  });
-  const { authorizationCode } = await client.authorize(authRequest);
-
-  const tokens = await fetchTokens(authRequest, authorizationCode);
-  await client.setTokens(tokens);
-
-  return tokens.access_token;
-}
-
-async function fetchTokens(authRequest: OAuth.AuthorizationRequest, authCode: string): Promise<OAuth.TokenResponse> {
-  const params = new URLSearchParams();
-  params.append("client_id", clientId);
-  params.append("code", authCode);
-  params.append("verifier", authRequest.codeVerifier);
-  params.append("grant_type", "authorization_code");
-  params.append("redirect_uri", authRequest.redirectURI);
-
-  const response = await fetch("https://oauth2.googleapis.com/token", { method: "POST", body: params });
-
-  if (!response.ok) {
-    const responseText = await response.text();
-    throw new Error(`Error while fetching tokens: ${response.status} (${response.statusText})\n${responseText}`);
-  }
-
-  return (await response.json()) as OAuth.TokenResponse;
-}
-
-async function refreshTokens(refreshToken: string): Promise<OAuth.TokenResponse> {
-  const params = new URLSearchParams();
-  params.append("client_id", clientId);
-  params.append("refresh_token", refreshToken);
-  params.append("grant_type", "refresh_token");
-
-  const response = await fetch("https://oauth2.googleapis.com/token", { method: "POST", body: params });
-
-  if (!response.ok) {
-    const responseText = await response.text();
-    // Throwing an error with the status text and including the response body
-    throw new Error(`Error while fetching tokens: ${response.status} (${response.statusText})\n${responseText}`);
-  }
-
-  const tokenResponse = (await response.json()) as OAuth.TokenResponse;
-  tokenResponse.refresh_token = tokenResponse.refresh_token ?? refreshToken;
-
-  return tokenResponse;
-}
-
-export async function getEmail(): Promise<string | undefined> {
-  const tokenSet = await client.getTokens();
-
-  const idToken = tokenSet?.idToken;
-  if (!idToken) {
-    return;
-  }
-
-  const { email } = jwt_decode<{ email?: string }>(idToken);
-  return email;
-}
