+++ extensions/google-workspace/src/create-google-document.tsx
@@ -1,5 +1,8 @@
 import { createDocFromUrl } from "./helpers/docs";
+import { withGoogleAuth } from "./components/withGoogleAuth";
 
-export default async function Command(props: { arguments: Arguments.CreateGoogleDocument }) {
+async function Command(props: { arguments: Arguments.CreateGoogleDocument }) {
   await createDocFromUrl("document", props.arguments.title);
 }
+
+export default withGoogleAuth(Command);
