+++ extensions/google-workspace/src/create-google-form.tsx
@@ -1,5 +1,8 @@
 import { createDocFromUrl } from "./helpers/docs";
+import { withGoogleAuth } from "./components/withGoogleAuth";
 
-export default async function Command(props: { arguments: Arguments.CreateGoogleForm }) {
+async function Command(props: { arguments: Arguments.CreateGoogleForm }) {
   await createDocFromUrl("forms", props.arguments.title);
 }
+
+export default withGoogleAuth(Command);
