+++ extensions/google-workspace/src/create-google-presentation.tsx
@@ -1,5 +1,8 @@
 import { createDocFromUrl } from "./helpers/docs";
+import { withGoogleAuth } from "./components/withGoogleAuth";
 
-export default async function Command(props: { arguments: Arguments.CreateGooglePresentation }) {
+async function Command(props: { arguments: Arguments.CreateGooglePresentation }) {
   await createDocFromUrl("presentation", props.arguments.title);
 }
+
+export default withGoogleAuth(Command);
